// types/data-products.ts
export type DataProducts = {
  ingredients: string[];
  salad: string[];
  soup: string[];
  meat: string[];
  garnish: string[];
  dessert: string[];
  meat_fish: string[];
  semifinished: string[];
  staff: string[];
  staff_first: string[];
  staff_garnish: string[];
  staff_main: string[];
  staff_snacks: string[];
};